a/b double

+++ c/les1/a+b/Program.cs
@@ -1,3 +1,3 @@
-﻿int a = 12;
-int b = 5;
+﻿double a = 12;
+double b = 5;
 Console.WriteLine(a/b);
